Create module which include main class methods.
